fix(keysign): address review on join verify scaffolding (#4549)
- Move 4 join-verify @Preview composables out of JoinKeysignView into
  their respective Verify*Screen.kt files (siblings of the existing
  initiator-path previews). Drops 10 model imports that were only
  pulled into JoinKeysignView for preview construction.
- Standardize the join-flow confirm CTA: all 4 variants now use
  verify_swap_sign_button ("Sign transaction") instead of Send shipping
  "Join Keysign" while Swap/Deposit/SignMessage shipped "Sign transaction".
- Rename showToolbar -> hasToolbar in VerifySwapScreen (signatures,
  internal forwards, V2Scaffold takeIf blocks, preview, and call sites
  in JoinKeysignView + PreviewActivity) so the Verify* family no longer
  hosts both spellings of the same concept.
- Remove defaults on hasToolbar / onBackClick in VerifySignMessageScreen
  so onBackClick can no longer be silently dropped when hasToolbar=false,
  matching VerifySwapScreen's required-params API. Update the only
  non-join caller and the existing preview to pass them explicitly.
