+++ app/src/main/java/com/vultisig/wallet/ui/screens/sign/VerifySignMessageScreen.kt
@@ -26,6 +26,7 @@ import com.vultisig.wallet.ui.components.buttons.VsButton
 import com.vultisig.wallet.ui.components.buttons.VsButtonVariant
 import com.vultisig.wallet.ui.components.launchBiometricPrompt
 import com.vultisig.wallet.ui.components.topbar.VsTopAppBar
+import com.vultisig.wallet.ui.models.sign.SignMessageTransactionUiModel
 import com.vultisig.wallet.ui.models.sign.VerifySignMessageUiModel
 import com.vultisig.wallet.ui.models.sign.VerifySignMessageViewModel
 import com.vultisig.wallet.ui.utils.asString
@@ -57,8 +58,10 @@ internal fun VerifySignMessageScreen(viewModel: VerifySignMessageViewModel = hil
 
     VerifySignMessageScreen(
         state = state,
+        hasToolbar = false,
         confirmTitle = stringResource(R.string.verify_swap_sign_button),
         onConfirm = viewModel::confirm,
+        onBackClick = {},
         onFastSignClick = {
             if (!viewModel.tryToFastSignWithPassword()) {
                 authorize()
@@ -70,11 +73,11 @@ internal fun VerifySignMessageScreen(viewModel: VerifySignMessageViewModel = hil
 @Composable
 internal fun VerifySignMessageScreen(
     state: VerifySignMessageUiModel,
+    hasToolbar: Boolean,
     confirmTitle: String,
     onFastSignClick: () -> Unit,
     onConfirm: () -> Unit,
-    hasToolbar: Boolean = false,
-    onBackClick: () -> Unit = {},
+    onBackClick: () -> Unit,
 ) {
     val transactionUiModel = state.model
     VerifySignMessageScreen(
@@ -94,11 +97,11 @@ private fun VerifySignMessageScreen(
     method: String,
     message: String,
     hasFastSign: Boolean,
+    hasToolbar: Boolean,
     confirmTitle: String,
     onFastSignClick: () -> Unit,
     onConfirm: () -> Unit,
-    hasToolbar: Boolean = false,
-    onBackClick: () -> Unit = {},
+    onBackClick: () -> Unit,
 ) {
     Scaffold(
         modifier = Modifier.fillMaxSize(),
@@ -165,6 +168,28 @@ private fun VerifySignMessageScreenPreview() {
         message = "message",
         confirmTitle = "Sign",
         hasFastSign = false,
+        hasToolbar = false,
+        onFastSignClick = {},
+        onConfirm = {},
+        onBackClick = {},
+    )
+}
+
+@Preview
+@Composable
+private fun JoinKeysignSignMessageVerifyPreview() {
+    VerifySignMessageScreen(
+        state =
+            VerifySignMessageUiModel(
+                model =
+                    SignMessageTransactionUiModel(
+                        method = "personal_sign",
+                        message = "Sign in to Uniswap",
+                    )
+            ),
+        hasToolbar = true,
+        confirmTitle = stringResource(R.string.verify_swap_sign_button),
+        onBackClick = {},
         onFastSignClick = {},
         onConfirm = {},
     )
